- callbacks in p2pquerier are now associated to a specific query - started to implement live pages for results. Still doesn't work, and lot   of ugly stuff.
This checkin should change a bit the results page presentation, but it
hopefully still works as before.

TODO now : move most of the SearchForm.child_search() content (especially
the local and distributed queries) into the ResultsPage.__init__() method
in order to prevent ugly and buggy checks like "if not words: ...".
Not sure it will solve the QueueUnderflow stuff, though